:sparkles: GroupNotificationDataSource 추상화
